requirements.txt
anthropic>=0.40.0
pdfplumber>=0.11.0
openpyxl>=3.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
duckduckgo-search>=6.0.0
